feat: add link endpoint and make endpoints strict

+++ monty-db/src/handler.ts
@@ -1,11 +1,11 @@
 import { HEADERS } from './headers'
 import { checkAuth } from './tokens'
-import { putKeys, getKeys, deleteKeys } from './keyspace'
+import { putKeys, getKeys, deleteKeys, listKeys } from './keyspace'
 
 export async function handleRequest(request: Request): Promise<Response> {
-  if (['POST', 'GET', 'DELETE', 'PUT'].indexOf(request.method) === -1) {
-    return new Response(null, { status: 405 })
-  }
+  // if (['POST', 'GET', 'DELETE', 'PUT'].indexOf(request.method) === -1) {
+  //   return new Response(null, { status: 405 })
+  // }
   // check the auth token
   const is_valid = await checkAuth(request)
   if (is_valid instanceof Response) {
@@ -18,15 +18,20 @@ export async function handleRequest(request: Request): Promise<Response> {
   }
 
   // Check the body is valid json and parse it
+  const path = new URL(request.url).pathname
   if (request.method === 'GET') {
-    return await getKeys(request)
+    if (path === '/get') {
+      return await getKeys(request)
+    } else if (path === '/list') {
+      return await listKeys(request)
+    }
   }
 
   let json: any
   try {
     // we're using this instead of Request.json() because that logs an error internally on fail
     const text = await request.text()
-    console.log(text)
+    // console.log(text)
     json = JSON.parse(text)
   } catch (e) {
     if (e instanceof SyntaxError) {
@@ -43,11 +48,11 @@ export async function handleRequest(request: Request): Promise<Response> {
   // console.log(JSON.stringify(json.config, null, 4))
 
   // actually execute the request
-  if (request.method === 'POST' || request.method === 'PUT') {
+  if (request.method === 'POST' || (request.method === 'PUT' && path === '/')) {
     return putKeys(request, json)
-  } else if (request.method === 'DELETE') {
+  } else if (request.method === 'DELETE' && path === '/') {
     return deleteKeys(request, json)
   } else {
-    throw new Error('unreachable')
+    return new Response(null, { status: 404 })
   }
 }
